Selectively disable image optimization
Images loaded from CDNs are not optimized on the application server.
In the game box this leads to a minor degredation im image quality and
the thumbnail shows better results because it is actually made for this
size.

All images served from the application server keep image optimization as
it can reduce bandwidth usage.