Various bug fixes and enhancements (#60)
* Add component for displaying errors

* Handle errors when sending friend rqeusts

* Add very basic error handling in multiple places

The error handling is in most cases unnecessarily verbose and there's a whole lot of repeated code. I have created #56 to enhance this at a later date.

* Move component definition outside render

* Upgrade verification token from unique to id

* Show my own collection status on profiles

Closes #59

* No error message in the log on profile pages

Closes #53

* Fix game page for games without image

Closes #54

* Do not show incomplete profiles

Closes #34

* Fix error on empty results in game search API

* Show feedback when search results are empty

Closes #57

* Do not show info for empty search terms

* Fix error messages for aborted search queries

* Memoize user data

This should increase page performance somewhat

* Sort games by player count in dashboard

Games with more players are now higher up. This takes a bit of pressure from #36 (but is still not nearly enough to fully solve that issue)