Remove linting warning
TanStack Query comes with linting rules that tell me to use all parameters used in the query function in the query key also. Sadly that doesn't work when including the parameter in a indirect way like I did. But I really like that the linter tries telling me I made a mistake so I restructured the code in a way that's a bit harder to maintain to make the linter work correctly.