API V2 (#100)
* Update first API to v2

* Do some stuff I want to get rid of before branching

* Move relationship API to Route Handler

* Move relationship API to Route Handler

* Implement new relationship normalization

This uses a much cleaner authorization check for fields that may be private.

* Add type removed by accident

* Remove impossible null value from type

* Fix data type in API route

* Use new relationship hooks with correct types

* Fix api bug

* Add new game API

* Prefetch game data on server

* Move game data hydration into component

* Fix build errors

* Remove unnecessary empty objects

* Use fetchGames when prefetching collections

This is slightly slower because not all games are fetched with the same query as the collection states but this way stale games get updated every time a collection is generated.

* Use game API v2 in profile editor

* Add hook to get game via query

This makes querying games much easier. I'm still not entirely satisfied with the design though.

* Simplify hook API

* Remove linting warning

TanStack Query comes with linting rules that tell me to use all parameters used in the query function in the query key also. Sadly that doesn't work when including the parameter in a indirect way like I did. But I really like that the linter tries telling me I made a mistake so I restructured the code in a way that's a bit harder to maintain to make the linter work correctly.

* Use cached data when available

* Use new game hooks in GamePill

* Remove old game data api

* Add DELETE method to relationship API

* Add testing library

* Add basic test cases to relation api v2

* Add GitHub action

* Rebuild package-lock

* Update workflow to use current action versions

* Downgrade next-auth to make it work in the tests

* Update test workflow to also run for pull requests

* Update workflow to only run on pull requests

Pushes without a pull request attached should never end up in main anyway.

* Rename server function to get feature flags

Starting the function with use implied it was a hook and the build failed because I can not use hooks on the server. It's not a hook, though, so I can just rename the function.

* Allow starting the test workflow manually

* Add more tests to relationship API

* Add reverse direction search

* Add tests for friend requests

* Add test to convertToUserProfile

* Only run tests matching the pattern I use

* Fix import of changed module

* Add basic tests for relationship DELETE API

* Add tests for relationship DELETE API

* Use relationship hook for sent friend requests

* Refresh package lock

* Add collection API

* Fix unreachable code

* Re-use utility function

* Remove error log

* Add useGameCollection as API hook

Data is already prefetched in the collection but not made available to the hook so a call per game is still necessary.

* Prefetch game collection data

* Fix removing last collection status from a game

* Fix sporadically failing test

* Remove unused old API routes

* Revert "Remove unused old API routes"

This reverts commit 3717dda1406eb29917c5d7909e4d5c9bc08fea65.

* Add API v2 function for receiving game collections

* Rename game collection to collection status

From now on a collection status is a users relation to a game and a game collection is the sum of all collection status of a user.

* Add hooks for using user collections

* Add keyboard handlers to game pills

* Optimize game collection prefetching