Add Mantine for modal handling
This makes building the modal much easier. I would have gone with Mantine when beginning the project now anyway. I hesitated switching now because it makes everything a bit more complex and is also bad for the user because of the increased bundle size. But now I reached a point where I think the added complexity is worth it and the bundle size isn't that much worse (eg. 11% increase when opening the collection) so I'll start gradually adapting Mantine.