Public profile page (#123)
* Add simple custom 404 page

* Add ability to completely hide ones profile

* Add background color to new profile header

Does not have to be set from the outside anymore.

* Fix BGG name showing empty when not set

* Move user profile to its own component

* Add mock function for server side i18n

* Add public user profile

* Add basic public profile page

* Update profile sharing url to public profile

* Reduce data collected for public profile page

* Add component for user cards

Can be used on public profiles to show profile information besides general information about Meeple Party and the login box. Some things are currently duplicated between this and the profile header.

* Simple code cleanup

* Move profile header username to own component

* Move local env file to .env

* Refresh package-lock

* Add component for real name in profile

* Reduce duplicated code

* Remove unused datatype

* Cleanup

* Update changed snapshot

* Add snapshot test for new component

* Add BGG Icon component

* Add component for additional profile information

* Use additional information component in structures

* Add tests for UserCard

* Remove duplicate place from user card

* Code cleanup

* Add component for logo

The static logo file in the public folder is still needed because it is utilized in the NextAuth theme that only supports static files. It should be safe to remove when doing #86.

* Move login button to its own component

* Design for public profile page

* Allow redirecting to specific pages on login

The redirect is currently still blocked by the custom auth page and needs to be passed to the signIn function there.

* Use callback url when signing in